feat(structure): move util package to pkg package
