+++ pkg/api/apps/logs.go
@@ -11,9 +11,9 @@ import (
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/gin-gonic/gin"
-	"github.com/hackclub/hack-as-a-service/pkg/api/util"
 	"github.com/hackclub/hack-as-a-service/pkg/db"
 	"github.com/hackclub/hack-as-a-service/pkg/dokku"
+	"github.com/hackclub/hack-as-a-service/pkg/util"
 )
 
 func handleGETLogs(c *gin.Context) {
