+++ pkg/api/builds/logs.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
-	"github.com/hackclub/hack-as-a-service/pkg/api/util"
 	"github.com/hackclub/hack-as-a-service/pkg/db"
 	"github.com/hackclub/hack-as-a-service/pkg/dokku"
+	"github.com/hackclub/hack-as-a-service/pkg/util"
 )
 
 func handleGETLogs(c *gin.Context) {
